read settings from environment variables instead of subbing them in during numbas-setup
fixes #1

It's not clear to me whether the migrate/collectstatic stuff in
numbas-setup.py gets run every time you run the container. Maybe
collectstatic should be part of the build step, and migrate doesn't do
anything if the database is already up to date.

I get some log messages like "'DEBUG' cast as 'None'" in the daphne
container, even though it seems to work. I'm not sure what that's about.