Move the setup stuff to explicit commands.
Rather than running the installation scripts whenever the whole recipe
runs, this removes the entrypoint from the numbas-setup container, so
you have to explicitly run the installation script with
`docker-compose run`.

The setup scripts are bound as a volume in the container, rather than
being copied into it.

This also adds a script `get_secret_key`, to generate the SECRET_KEY=
line.